실시간 코로나 확진자 리스트를 가져오는 부분을 커스텀 훅으로 만들어서 관리합니다
