+++ src/interfaces/realTimeConfirmedCase.ts
@@ -1,0 +1,6 @@
+export default interface realTimeConfirmedCase {
+  id: number
+  time: string
+  province: string
+  confirmedCaseCount: number
+}
